requirements.txt
astroid==1.6.5
autopep8==1.4.3
backports.functools-lru-cache==1.5
configparser==3.7.1
Django==1.11
djangorestframework==3.9.1
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pylint==1.9.4
pytz==2018.9
singledispatch==3.4.0.3
six==1.12.0
wrapt==1.11.1
